+++ AddOns/BetterBlizzFrames/temp_era/gui.lua
@@ -2175,7 +2175,7 @@ local function CreateSimpleDropdown(name, parentFrame, labelText, settingKey, op
             local displayText = GetDisplayText(option)
             -- Create each item as a button
             local button = rootDescription:CreateButton(displayText, function()
-                BetterBlizzFramesDB[settingKey] = option
+                BetterBlizzFramesDB[settingKey] = tonumber(option) or option
                 dropdown:SetDefaultText(displayText)
                 if toggleFunc then
                     toggleFunc(option)