+++ AddOns/BetterBlizzFrames/temp_era/modules/nameFix.lua
@@ -398,6 +398,8 @@ C_Timer.After(1, function()
 end)
 
 local function SetPartyFont(font, size, outline, size2)
+    size = tonumber(size)
+    size2 = tonumber(size2)
     for i = 1, 5 do
         local frame = _G["CompactPartyFrameMember"..i]
         if frame then
@@ -438,9 +440,10 @@ local function SetPartyFont(font, size, outline, size2)
 end
 
 local function SetUnitFramesFont(font, size, outline)
+    size = tonumber(size)
     local anyFailed = false
     for _, frame in ipairs(frames) do
-        local newSize = tonumber(size)
+        local newSize = size
         if frame == PetFrame or frame == TargetFrameToT then
             if newSize >= 13 then
                 newSize = newSize - 3