+++ AddOns/BetterBlizzFrames/CHANGELOG.md
@@ -1,3 +1,47 @@
+# BetterBlizzFrames 1.9.9f
+## Midnight
+### Bugfix
+- Fix lua error from instant combo points setting after changes to PRD from Blizzard
+
+# BetterBlizzFrames 1.9.9e
+## Midnight
+### Bugfix
+- Even more fixes for font settings for 12.0.7
+
+# BetterBlizzFrames 1.9.9d
+## Midnight
+### Tweak
+- WoW 12.0.7 comes with more bugs. Important aura filter is broken because of Blizzard and shows a bunch of trash auras. Temporarily disabled important auras until Blizzard fixes it.
+
+# BetterBlizzFrames 1.9.9c
+## Midnight
+### Bugfix
+- More fixes for font settings for 12.0.7
+
+# BetterBlizzFrames 1.9.9b
+## Midnight
+### New
+- New JFarm profile (www.twitch.tv/jfarm_ & www.youtube.com/@Jfarm_). Thank you for sharing.
+### Bugfix
+- Fix lua error in unitframe font change setting
+
+# BetterBlizzFrames 1.9.9
+## Mists of Pandaria
+### New
+- Add Edit Mode transparency setting (Misc & in Edit Mode menu)
+- Add PartyFrame range alpha setting (General)
+- Add Improved AddonList setting (Misc)
+### Bugfix
+- Fix Target/FocusFrame background texture being too low on Minus mobs
+- Fix adding duration to smoke bomb aura for MoP.
+- Fix PlayerFrame elite texture being wrong size after MoP changes.
+## Midnight
+### Bugfix
+- Fix "Hide Player Resource/Power" hiding some classes resources despite ignore setting being enabled.
+## Classic Era
+### Bugfix
+- Fix Player Castbar not getting moved with settings after removing it for MoP & TBC.
+
 # BetterBlizzFrames 1.9.8e
 ## Mists of Pandaria
 ### Bugfix